Moved all variables with texts to separate files in the constants folder.
