+++ src/constants/exercisesTableTexts.ts
@@ -1,0 +1,3 @@
+export const text = 'Czy na pewno chcesz usunąć to ćwiczenie?';
+
+export const textInformation = 'Należy wypełnić wszystkie pola dotyczące ćwiczenia!'
