+++ src/constants/partsOfPlanTableTexts.ts
@@ -1,0 +1,3 @@
+export const text = 'Czy na pewno chcesz usunąć tę część planu? Spowoduje to także usunięcie wszystkich ćwiczeń przypisanych do tej części planu';
+
+export const textInformation = 'Należy podać nazwę części planu!'
