+++ src/constants/plansListTexts.ts
@@ -1,0 +1,3 @@
+export const text = 'Czy na pewno chcesz usunąć ten plan? Spowoduje to także usunięcie wszystkich części planu przypisanych do tego planu';
+
+export const textInformation = 'Należy podać nazwę planu!'
