+++ src/constants/rulesTableTexts.ts
@@ -1,0 +1,3 @@
+export const text = 'Czy na pewno chcesz usunąć tę zasadę progresji?'
+
+export const textInformation = 'Należy podać treść zasady progresji!'
